Extract web related types to 'Web/Types'.
